+++ src/event/mod.rs
@@ -7,6 +7,7 @@ pub enum Event {
     Key(KeyEvent),
     Action(Action),
     MouseDown(WidgetId),
+    MouseUp,
     Tick(u64),
     Resize(u16, u16),
 }